Update Mashup styling.
Update the Mashup page with styling that addresses the redundant miniboxes between search results, redresses the No Results Found message to be styled as other search results, and changes the color and increases the font size of the search query fields. Add spacing between the search field titles and search field queries.